- Fixed various bugs - Added Full Integration Tests - Bumped dependency version of System.Text.Json to 6.0.0 as 5.0.0 have know dependency vulnerabilities and 6.0.0 is also the LTS version

+++ TrelloDotNet/TrelloDotNet/Model/Checklist.cs
@@ -29,13 +29,13 @@ namespace TrelloDotNet.Model
         [JsonInclude]
         public string CardId { get; private set; }
 
-        //todo - Add Support for position (research in order to understand)
-        /*
+        /// <summary>
+        /// The position of the Checklist on the card
+        /// </summary>
         [JsonPropertyName("pos")]
         [QueryParameter]
-        public int Position { get; set; }
-        */
-
+        public decimal Position { get; set; }
+        
         /// <summary>
         /// The items of the Checklist
         /// </summary>
